backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-auditlog==3.0.0
django-cors-headers==4.7.0
django-debug-toolbar==4.4.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
paramiko==3.5.1
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
